SO-53 Refactoring
- README updates
- Assets added for readme
- Fixed layout for side navbar functionality
