+++ README.md
@@ -1,4 +1,109 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+# SacOverflow
+
+Welcome to Sacoverflow's GitHub! 🚀 This project is aimed at assisting Project Managers in overviewing their organization's finances and overall management about projects and tasks within their org. The SacOverflow development team, consisting of Computer Science undergraduates from California State University, Sacramento, is behind this project.
+
+This project is aimed at assisting Project Managers in managing their team and projects. By making use of this software, supervisors or relevant stakeholders will be allowed to make more informed decisions within their organization and simplify their day-to-day task duties. This software is to be developed with the intention of being used on both Mobile and Web based platforms.
+
+[![MIT License][MIT-shield]](https://opensource.org/licenses/MIT)
+
+[![NextJS Icon][nextjs-shield]](https://nextjs.org/)
+[![Contributors][contributors-shield]][contributors-url]
+[![Forks][forks-shield]][forks-url]
+[![Stargazers][stars-shield]][stars-url]
+[![License][MIT-shield]](https://opensource.org/licenses/MIT)
+
+### Built With
+
+-   [![NextJS Icon][nextjs-shield]](https://nextjs.org/)
+-   [![ReactJS Icon](https://img.shields.io/badge/React-20232A?style=for-the-badge&logo=react&logoColor=61DAFB)](https://react.dev/)
+-   [![TypeScript Icon](https://img.shields.io/badge/TypeScript-007ACC?style=for-the-badge&logo=typescript&logoColor=white)](https://www.typescriptlang.org/)
+-   [![TailwindCSS Icon](https://img.shields.io/badge/Tailwind_CSS-38B2AC?style=for-the-badge&logo=tailwind-css&logoColor=white)](https://tailwindcss.com/)
+-   [![SupaBase Icon](https://img.shields.io/badge/Supabase-181818?style=for-the-badge&logo=supabase&logoColor=white)](https://supabase.com/)
+-   [![PostgreSQL Icon](https://img.shields.io/badge/PostgreSQL-316192?style=for-the-badge&logo=postgresql&logoColor=white)](https://www.postgresql.org/)
+-   [![Vercel Icon](https://img.shields.io/badge/Vercel-000000?style=for-the-badge&logo=vercel&logoColor=white)](https://vercel.com/)
+
+### Installation
+
+1. Clone the repository to your local machine and change directories into it. Stable builds will be based off the main branch.
+
+```bash
+git clone https://github.com/hashemJaber/SacOverflow.git
+cd SacOverflow
+```
+
+2. Install the dependencies using whatever package manager you prefer.
+
+```bash
+npm install
+# or
+yarn install
+# or
+pnpm install
+# or
+bun install
+```
+
+3. Run the development server.
+
+```bash
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
+```
+
+### Setup
+
+1. Create a `.env.local` file in the root directory of the project.
+2. Development phases there is a Mockaroo API key that is needed to be able to run the project. Please contact the lead for the key.
+3. Add the following to the `.env.local` file.
+
+```bash
+MOCKAROO_API_KEY=YOUR_API_KEY
+```
+
+### Features
+
+-   User Authentication (Login/Sign Up)
+-   Business Dashboard (Overview of finances and projects)
+-   Project Dashboard (Overview of project details and tasks)
+-
+
+### Mockup Designs
+
+-   <h5>Home Page</h5>
+      <p align='center'>
+          <img src="assets/Home Page.png" alt="Web Home" width="250" height="250">
+      </p>
+-   <h5>Login Page<h5/>
+      <p align='center'>
+          <img src="assets/Web Login Page.png" alt="Web Home" width="250" height="250">
+      </p>
+-   <h5>Sign Up Page<h5/>
+      <p align='center'>
+          <img src="assets/Web Sign Up Page.png" alt="Web Home" width="250" height="250">
+      </p>
+-   <h5>Dashboard Page<h5/>
+      <p align='center'>
+          <img src="assets/Web Dashboard.png" alt="Web Home" width="250" height="250">
+      </p>
+-   <h5>Projects Dashboard Page<h5/>
+      <p align='center'>
+          <img src="assets/Web Projects Page.png" alt="Web Home" width="250" height="250">
+      </p>
+-   <h5>Projects Details Dashboard Page<h5/>
+      <p align='center'>
+          <img src="assets/Web Project Detail Page.png" alt="Web Home" width="250" height="250">
+      </p>
+
+### API Reference
+
+-   TBD
+
+    This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
 ## Getting Started
 
@@ -24,8 +129,8 @@ This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-opti
 
 To learn more about Next.js, take a look at the following resources:
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+-   [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+-   [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
 
 You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
 
@@ -34,3 +139,15 @@ You can check out [the Next.js GitHub repository](https://github.com/vercel/next
 The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
 Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+
+<!-- MARKDOWN LINKS & IMAGES -->
+<!-- https://www.markdownguide.org/basic-syntax/#reference-style-links -->
+
+[contributors-shield]: https://img.shields.io/github/contributors/othneildrew/Best-README-Template.svg?style=for-the-badge
+[MIT-shield]: https://img.shields.io/badge/license-MIT-blue.svg
+[nextjs-shield]: https://img.shields.io/badge/next.js-000000?style=for-the-badge&logo=nextdotjs&logoColor=white
+[contributors-url]: https://github.com/hashemJaber/SacOverflow/graphs/contributors
+[forks-shield]: https://img.shields.io/github/forks/othneildrew/Best-README-Template.svg?style=for-the-badge
+[forks-url]: https://github.com/hashemJaber/SacOverflow/network/members
+[stars-shield]: https://img.shields.io/github/stars/othneildrew/Best-README-Template.svg?style=for-the-badge
+[stars-url]: https://github.com/hashemJaber/SacOverflow/stargazers
