+++ src/app/(dashboard)/layout.tsx
@@ -1,0 +1,44 @@
+'use client';
+import SideNavbar from '@/components/SharedComponents/SideNavbar';
+import { useState } from 'react';
+
+// import css
+import './base.css';
+
+const Layout = ({ children }) => {
+    // for opening and closing sidebar
+    const [showSidebar, setShowSidebar] = useState(false);
+
+    const Hamburger = ({ setter }) => {
+        return (
+            <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+                strokeWidth="1.5"
+                stroke="currentColor"
+                className="overlay-icon"
+                onClick={() => {
+                    setter((oldVal) => !oldVal);
+                }}
+            >
+                <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    d="M3.75 6A2.25 2.25 0 016 3.75h2.25A2.25 2.25 0 0110.5 6v2.25a2.25 2.25 0 01-2.25 2.25H6a2.25 2.25 0 01-2.25-2.25V6zM3.75 15.75A2.25 2.25 0 016 13.5h2.25a2.25 2.25 0 012.25 2.25V18a2.25 2.25 0 01-2.25 2.25H6A2.25 2.25 0 013.75 18v-2.25zM13.5 6a2.25 2.25 0 012.25-2.25H18A2.25 2.25 0 0120.25 6v2.25A2.25 2.25 0 0118 10.5h-2.25a2.25 2.25 0 01-2.25-2.25V6zM13.5 15.75a2.25 2.25 0 012.25-2.25H18a2.25 2.25 0 012.25 2.25V18A2.25 2.25 0 0118 20.25h-2.25A2.25 2.25 0 0113.5 18v-2.25z"
+                />
+            </svg>
+        );
+    };
+
+    return (
+        <main className="flex flex-col md:flex-row relative">
+            {/* Hamburger */}
+            <Hamburger setter={setShowSidebar} />
+            <SideNavbar show={showSidebar} setter={setShowSidebar} />
+            {children}
+        </main>
+    );
+};
+
+export default Layout;
