--- src/app/layout.js
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
-
-export default function RootLayout({ children }) {
- return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  )
-}
