+++ src/components/SharedComponents/SideNavbar.tsx
@@ -29,26 +29,6 @@ const NavigationItems = [
         path: '/dashboard',
     },
     {
-        name: 'Transactions',
-        icon: (
-            <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth={1.5}
-                stroke="currentColor"
-                className="navitem-icon"
-            >
-                <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m0 12.75h7.5m-7.5 3H12M10.5 2.25H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z"
-                />
-            </svg>
-        ),
-        path: '/transactions',
-    },
-    {
         name: 'Invoices',
         icon: (
             <svg
@@ -162,7 +142,6 @@ const AccountControlItems = [
 const SideNavbar = ({ show, setter }) => {
     const pathname = usePathname();
 
-    // const appendShown = show ? 'ml-0' : 'ml-[-250px] md:ml-0';
     const appendShown = show ? 'disabled' : '';
 
     // TODO: implement modal overlay so if user clicks outside of sidebar, it closes
